Start to abstract away tiles in their own components
